requirements.txt
beautifulsoup4==4.12.2
django_haystack==3.2.1
pandas==1.5.3
PyPDF2==3.0.1
regex==2023.6.3
requests==2.28.2
streamlit==1.19.0
streamlit_lottie==0.0.5
